Adds about.html page with list of project contributors
Adds make_about.py, which retrieves json object from api.github.com with list of contributors, saves it as a dict, and passes it to about.html template which displays it as an html list element. Changes base.html template’s `<address>` element to link to the new page. Adds call to python make_about.py to make.sh.